requirements.txt
decorator==4.1.2
Django==1.11.3
django-debug-toolbar==1.8
django-extensions==1.8.1
flake8==3.4.1
importmagic==0.1.7
ipython==4.2.1
ipython-genutils==0.2.0
jedi==0.10.2
mccabe==0.6.1
pexpect==4.2.1
pickleshare==0.7.4
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.5.0
pytz==2017.2
rope==0.10.5
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
yapf==0.16.3
